--- src/components/Filter.tsx
@@ -1,63 +1,0 @@
-import React from 'react';
-import classNames from 'classnames';
-import { Todo } from '../types/Todo';
-import { Filter } from '../types/Filter';
-
-interface Props {
-  todos: Todo[],
-  filter: Filter,
-  setFilter: (filter: Filter) => void;
-}
-
-export const FilterTodo: React.FC<Props> = ({ todos, filter, setFilter }) => (
-  <footer className="todoapp__footer" data-cy="Footer">
-    <span className="todo-count" data-cy="todosCounter">
-      {`${todos.length} items left`}
-    </span>
-
-    <nav className="filter" data-cy="Filter">
-      <a
-        data-cy="FilterLinkAll"
-        href="#/"
-        className={classNames(
-          'filter__link',
-          { selected: filter === Filter.All },
-        )}
-        onClick={() => setFilter(Filter.All)}
-      >
-        All
-      </a>
-
-      <a
-        data-cy="FilterLinkActive"
-        href="#/active"
-        className={classNames(
-          'filter__link',
-          { selected: filter === Filter.Active },
-        )}
-        onClick={() => setFilter(Filter.Active)}
-      >
-        Active
-      </a>
-      <a
-        data-cy="FilterLinkCompleted"
-        href="#/completed"
-        className={classNames(
-          'filter__link',
-          { selected: filter === Filter.Completed },
-        )}
-        onClick={() => setFilter(Filter.Completed)}
-      >
-        Completed
-      </a>
-    </nav>
-
-    <button
-      data-cy="ClearCompletedButton"
-      type="button"
-      className="todoapp__clear-completed"
-    >
-      Clear completed
-    </button>
-  </footer>
-);
